requirements.txt
torch
torchvision
pandas
transformers
Pillow
tqdm
accelerate
datasets 
albumentations 
numpy
scipy 
bitsandbytes 